Move rendering process configuration from GuiPanel to KarelWorld to avoid NPE
